refactor: login authencitaionprincipal로 수정 & testcontroller작성

+++ src/main/java/com/umc/linkyou/web/dto/UserResponseDTO.java
@@ -13,7 +13,7 @@ public class UserResponseDTO {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class JoinResultDTO{
-        Integer userId;
+        Long userId;
         LocalDateTime createdAt;
     }
 
@@ -22,7 +22,7 @@ public class UserResponseDTO {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class LoginResultDTO{
-        Integer userId;
+        Long userId;
         String accessToken;
     }
